refactor(effects): extract the shared EffectEngine; Store delegates to it
Pulls the scheduling engine out of `Store` into a single `EffectEngine<Action>`
that both `Store` and (next) `TestStore` drive — one implementation, zero
duplication. The engine owns the registries and honours every component's
`EffectScheduling` (debounce/throttle/replace/delay/cancel and the channel path);
the two substrate differences are injected:

- `clock` — `ContinuousClock` in production; `ImmediateClock`/`TestClock` in tests.
- `send` — where produced actions go (the Store's `@MainActor` process hop).

It is a `@MainActor final class` because its delay timers re-enter and mutate the
registries on fire — unlike the 1a `EffectScheduler` (deleted as a redundant
Store-only class), this engine is the single shared scheduler, so the reference
type is justified, not redundant. `Store` now holds a `lazy var engine` (lazy so
its `send` can capture the fully-initialised Store) and resolves the clock from the
environment at init.

Behaviour-preserving: 426 tests green (--enable-all-traits), lint clean, DocC 0
warnings. Sets up 1b′ (TestStore on the same engine).
